bsp: add initial buildroot framework and defconfig
Add the initial framework of folders and files to use this with
buildroot as a BR2_EXTERNAL location. Add a defconfig for the
BeagleBone Black that is initially a copy of the beaglebone_defconfig.

board/ and package/ directories will appear later once I have files to
add to these directories.

Signed-off-by: Joel Carlson <JoelsonCarl@gmail.com>